[refactor] feed domain
